+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   
 * Added vendor prefixes: `MESHOPT`, `POLUTROPON`, and `AGT` (#119).
 
+* Added vendor prefixes: `ALCM` and `SKYLINE`.
+
 * Extension names are now validated to have an upper-case prefix separated from the rest on the name with underscore (new `INVALID_EXTENSION_NAME_FORMAT` warning). 
 
 * Referencing an image of type defined by an extension from the core objects now results in `TEXTURE_INVALID_IMAGE_MIME_TYPE` error.