+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 * Added vendor prefixes: `ALCM` and `SKYLINE`.
 
-* Added `FOXIT` vendor prefix.
+* Added vendor prefixes: `FOXIT`, `KDAB`, and `CAPTURE`.
 
 * Extension names are now validated to have an upper-case prefix separated from the rest on the name with underscore (new `INVALID_EXTENSION_NAME_FORMAT` warning).